+++ CHANGELOG.md
@@ -22,6 +22,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `heights` flag
 - Added `selectors.specificity.highest10` metric
 - Added `selectors.maxPerRuleList` metric
+- Added `keyframes.definedAnimations` metric
+- Added `animations.withoutDefinitions` metric
 
 ### Changed